changed the event handling.
the event handling in the fmu class now notes during integration if an state event happend and also the time of the event. thus the incrementalfmu class can restart the integration from the last step of the lookahead until the event.

because IncrementalFMU::handleEvent now sets the time back and integrates until the state event, that should not but might cause a different behavior from before if you have overwritten the function and call it after your own event handling.

if this causes problems, it might make sense to differentiate between internally und externally caused events...