Fixed bug resulting in a segmentation fault
The result of the private function loadDLL was not checked and the bare
FMU was added without previously loading the function pointers.
Additionally some debug information was added in a loadDLL error message.
Because the export part of the fmipp project currently doesn't compile properly on my system,
the test cases were not executed. (Sorry!)