make bdf integrate the dymola model correctly
a big rework on the bdf stepper has been made in order to pass a testmodel generated by dymola. The changes were inspired by the 2.0 brach wich
has a very different structure than the current one. Therefore, the FMUModelexchange has been directly acessed by the stepper. this should not be
the standard in future versions of the steppers.