removed unnecessary memeber functions in class Integrator
since the sundials stepper now have direct acess to FMUModelExchangeBase, they do not need the methods from integrator anymore

note that the logic for these steppers is different for the 2.0 branch: In the other branch, the integrator gets a DynamicalSystem
object passed in the constructor. DynamicalSystem is a base class which inherits the functions setTime, setStates, getDerivatives,...
to FMUModelexchangeBase and FMU2ModelExchangeBase.