DynamicalSystem Class
First commit for importing the 2.0 brach:

The class DynamicalSystem will later be used to determine which parts of FMUModelExchangeBase are exposed to the integrator class and
which are not. The main idea is "keep eventhandling ut of the integration" meaning that the job of an integrator is to locate events,
not handle them.