Add FMUModelExchange_v2.cpp and .h
Class for FMUs for ModelExchage according to the 2.0 standard. Jacobians are either calculated via getDirectionalDerivative or with numeric differentiation
(done by the DynamicalSystem class) depending on the flag providesDirectionalDerivatives in the Modeldescription.

The new class iherits from FMUModelExchangebase and therefore has to convert between types of the fmi1.0 specification
and types from the 2.0 specification.

FMUMEs of type 1.0 and 2.0 have a common base class called dynamical system. This class serves as input for the
Integrator class.

Note that the numerical derivative is always used for FMUs of type 1.0 when getJac is called.