New initialization logic for class Integrator
the constructor of the Integrator class gets called inside the constructor of DynamicalSystem. This means, that the this pointer from
the FMU for modelexchange does not have to be used (safety). The initialize command should be used after nStates and nEventinds are
defined. The stepper cannow  be changed during simulation with Integrator::setType (currently not available in the API of FMUME).