apply tolerances for the steppers
the constructors of the different steppers have been rewritten, so the abstol and retol fields of Integrator::Properties are used if
the steppers allow this. For the other steppers (like euler), the given tolerances are ignored and set to infinity (the properties
object is passed as reference ).

after construction of the stepper (which happens in the constructor of FMUME), the tolerances can be obtained by using

   FMUMEBase::getIntegratorProperties().abstol
   FMUMEBase::getIntegratorProperties().reltol

By default, the toleraces specified by the defaultExperiment (in the modelDescription) are used.