Rename FMUME::getJacobian -> FMUME::getJac, workaround for Dymola bug
the function FMUModelExchange::getJacobian has been renamed to getJac so it properly overwrites the function defined in DnamicalSystem. This gives
the steppers acess to this functions.

Sundials is now informed in case the calculation of the Jacobian goes wrong. This feature is untested tough.

While testing it turned out that FMUs exported by dymola behave oddly when calculating Jacobians. This is probably the same bug as in the thread
https://bitbucket.org/siani/javafmi/issue/28/accessgetdirectionalderivatives-wrong. A workaround has been made. This workaround is disabled by
default and can be enabled by setting the enviroment variable DYMOLA2015_WORKAROUND or using the CMAKE GUI. The workaround has been tested on
two Dymola FMUs and worked very well for these examples. Setting the flag DYMOLA2015_WORKAROUND to true should not affect the functionality of
non-Dymola FMUs. The auto test cases pass with the workaround enabled.

Interestingly enough, the getJacobian Function is still very slow when applied to dymola models provided the FMU has many states. This problem will
be adressed in the near future - hopefully, the bad performance is not a problem with the FMus but with the wrapper function getJac...