implement completedIntegratorStep for all test FMUs
so far, step events were not tested extensively. Therefore the completedIntegratorStep was implemented in a trivial way for some FMUs.

Unfortunately, this means that the call to

   fmiCompletedIntegratorStep( ..., fmiBoolean* stepEventFlag )

returned a random flag leading to some failing test cases. This should be fixed now. Note that this commit is a quickfix and not a
refactoring (all of the cahnged FMUs do not contain step Events and the event flags have just been set to fmiFalse/fmi2False  ).