test cases for boolean-setters/boolean-getters
the FMUSDK example values is used to test the getters and setters for different value types. The tests for the boolean-getters is very extensive since
there was a bug with those functions in the past (see commit message of 6d3e3d5592e7a1a88cb849b9974e0766d515b8ca for details )

Interestingly enough, the continuous states are not integrated correctly while all other variables seem to behave as expected.